Added 'noslidenumbers' to "Package Options"
I added the 'noslidenumbers' option to the "Package Options" section with a
description of what it does.

Additionally, at the top of the "Package Options" I put how to call theme
options. When first working with mtheme, I struggled to find where to put the
package options, so I figured new users would find it helpful to have in the
ReadMe. It may be unnecessary since it's like every other Beamer theme in that
regard. Still, since it's a non-standard theme, I think it's good for a new user
to be told explicitly.