Add Makefile target to build the manual
Running the PyTeX interpreter does not really work. Moreover, I don't think it's
very good idea to depend on requests for such a simple task and to get the
contributor list in the document in the first place. I guess maintaining the
list manually is giving less headaches in the long run.