new frame title format implementation
As also suggested on
[stackexchange](http://tex.stackexchange.com/questions/112526/errors-wit
h-makeuppercase-and-beamers-insertsectionhead) the new implementation
patches `\sectionentry`, `\beamer@section` and `\beamer@@frametitle`
instead of expanding them.

In the process I also added three more hooks:

- `mthemesectiontitleformat`
- `mthemeframetitleformat`
- `mthemeplaintitleformat`

So the user can change the behavior of each element independently. As
default the new formats are dependent on `mthemetitleformat`.

For testing purposes I integrated @elliottd s frames from this PR #94

This should fix the `allowframebreaks` issue for real now.