Defined a length specific to frame title format
Used the \heightof{} command from the calc package to ensure that the
height used as a strut in the beamercolorbox is consistent.

The height is set to the height of the tallest character of the font in
a typeface independent manner, and also depending on whether the frame
title is set in small caps uppercase or regular case. This ensure that
the colorbox is always the height of the text plus 1.5ex above and
below the text independent of format, text size, ratio and typeface.