Added a second \@metropolis@frametitlestrut
I’ve put a `\@metropolis@frametitlestrut` at the beginning and the end
because of what I can see will be an error for multiline titles.
Perhaps what @benjamin-weiss was alluding too?

If we were to put the strut at the end of a multiline title, then if
the first line was nothing but lower case characters (for some reason…)
in a `frametitleformat=regular` situation then the strut would make
sure that the minimum height would still be that of a capital letter.
But, since this is a multiline title that only happens for the last
line.

If we thing of the opposite situation; the strut is at the beginning
but our final line is nothing but lowercase. Then the strut again makes
sure that the minimum height is that of a capital letter. But, this is
only applied to the first line now. To rectify this I’ve put a strut at
both the beginning and the end. Now, this won’t help If the presenter
has a multiline title that is three or more lines long (god forbid) but
in that case the standard leading and `\baselinskip` should take care
of it. Also, the central line(s) shouldn’t effect the padding around
the title anyway.