Fix known issue with beamercolortheme fly
Previously `\plain` slides were broken if the user overrode the colour theme with `fly`, as that theme does not define a background colour for `palette primary`. The solution checks whether `palette primary.bg` is undefined; if so, it automatically picks a colour complementary to `palette primary.fg` for the background of the plain slides.